started splitting up monolithic controller into seperate, more distinct controllers/files
