+++ controllers/test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/astaxie/beego"
-
 	"github.com/microsoft/mouselog/detect"
 	"github.com/microsoft/mouselog/trace"
 )
@@ -14,36 +13,10 @@ type APIController struct {
 	beego.Controller
 }
 
-var sessions map[string]*trace.Session
-
-func init() {
-	sessions = map[string]*trace.Session{}
-}
-
-// Session either returns an already existing session or creates and returns a new one.
-// If a new session has been created, the returned boolean will be true.
-func Session(sessionID string) (*trace.Session, bool) {
-	if val, ok := sessions[sessionID]; ok {
-		return val, false
-	}
-
-	sessions[sessionID] = trace.NewSession(sessionID)
-	return sessions[sessionID], true
-}
-
 type response struct {
 	Status string      `json:"status"`
 	Msg    string      `json:"msg"`
 	Data   interface{} `json:"data"`
-}
-
-func (c *APIController) SessionID() {
-	sessionID := c.StartSession().SessionID()
-
-	trace.AddSession(sessionID, c.Input().Get("websiteId"), c.Ctx.Input.UserAgent(), c.Ctx.Input.IP())
-
-	c.Data["json"] = sessionID
-	c.ServeJSON()
 }
 
 func (c *APIController) UploadTrace() {
